src/auth/auth.controller.ts
import { Body, Controller, Get, Post } from '@nestjs/common';
import { LoginUserDto } from './dto/login-user.dto';
import { RegisterUserDto } from './dto/register-user.dto';
import { AuthService } from './auth.service';

@Controller('auth')
export class AuthController {
  constructor(private authService: AuthService) {}

  @Post('/registeruser')
  registerUser(@Body() resgisterUserDto: RegisterUserDto) {
    return this.authService.registerUser(resgisterUserDto);
  }

  @Post('/loginuser')
  loginUser(@Body() loginUserDto: LoginUserDto) {
    return this.authService.loginUser(loginUserDto);
  }
}
